Add DMs implementation

+++ src/main/java/app/foxochat/service/impl/ChannelServiceImpl.java
@@ -14,14 +14,12 @@ import app.foxochat.exception.media.UploadFailedException;
 import app.foxochat.exception.member.MemberAlreadyInChannelException;
 import app.foxochat.exception.member.MemberInChannelNotFoundException;
 import app.foxochat.exception.member.MissingPermissionsException;
+import app.foxochat.exception.user.UserNotFoundException;
 import app.foxochat.model.Channel;
 import app.foxochat.model.Member;
 import app.foxochat.model.User;
 import app.foxochat.repository.ChannelRepository;
-import app.foxochat.service.ChannelService;
-import app.foxochat.service.GatewayService;
-import app.foxochat.service.MediaService;
-import app.foxochat.service.MemberService;
+import app.foxochat.service.*;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -44,15 +42,18 @@ public class ChannelServiceImpl implements ChannelService {
 
 	private final MediaService mediaService;
 
-	public ChannelServiceImpl(ChannelRepository channelRepository, MemberService memberService, @Lazy GatewayService gatewayService, MediaService mediaService) {
+	private final UserService userService;
+
+	public ChannelServiceImpl(ChannelRepository channelRepository, MemberService memberService, @Lazy GatewayService gatewayService, MediaService mediaService, UserService userService) {
 		this.channelRepository = channelRepository;
 		this.memberService = memberService;
 		this.gatewayService = gatewayService;
 		this.mediaService = mediaService;
+		this.userService = userService;
 	}
 
 	@Override
-	public Channel add(User user, ChannelCreateDTO body) throws ChannelAlreadyExistException {
+	public Channel add(User user, long partnerId, ChannelCreateDTO body) throws ChannelAlreadyExistException, UserNotFoundException {
 		Channel channel;
 
 		long isPublic = 0;
@@ -60,14 +61,26 @@ public class ChannelServiceImpl implements ChannelService {
 		if (body.isPublic() && body.getType() != ChannelConstant.Type.DM.getType())
 			isPublic = ChannelConstant.Flags.PUBLIC.getBit();
 
+		boolean isDM = partnerId != 0 && body.getType() == ChannelConstant.Type.DM.getType();
+
+		User partner = null;
+		if (isDM) {
+			partner = userService.getById(partnerId).orElseThrow(UserNotFoundException::new);
+		}
+
 		try {
-			channel = new Channel(body.getDisplayName(), body.getName(), isPublic, body.getType(), user);
+			channel = new Channel(body.getDisplayName(), body.getName(), isPublic, body.getType());
 			channelRepository.save(channel);
 		} catch (DataIntegrityViolationException e) {
 			throw new ChannelAlreadyExistException();
 		}
 
-		Member member = new Member(user, channel, MemberConstant.Permissions.ADMIN.getBit());
+		if (isDM) {
+			Member partnerMember = new Member(partner, channel, MemberConstant.Permissions.OWNER.getBit());
+			memberService.add(partnerMember);
+		}
+
+		Member member = new Member(user, channel, MemberConstant.Permissions.OWNER.getBit());
 		memberService.add(member);
 
 		log.debug("Channel ({}) by user ({}) created successfully", channel.getName(), user.getUsername());
@@ -88,8 +101,11 @@ public class ChannelServiceImpl implements ChannelService {
 
 	@Override
 	public Channel update(Member member, Channel channel, ChannelEditDTO body) throws Exception {
-		if (!member.hasAnyPermission(MemberConstant.Permissions.ADMIN, MemberConstant.Permissions.MANAGE_MESSAGES))
+		if (!member.hasAnyPermission(MemberConstant.Permissions.OWNER, MemberConstant.Permissions.ADMIN, MemberConstant.Permissions.MANAGE_MESSAGES))
 			throw new MissingPermissionsException();
+
+		if (channel.getType() == ChannelConstant.Type.DM.getType())
+			throw new ChannelNotFoundException();
 
 		try {
 			String name = body.getName();
@@ -110,7 +126,7 @@ public class ChannelServiceImpl implements ChannelService {
 			throw new UploadFailedException();
 		}
 
-		gatewayService.sendMessageToSpecificSessions(getRecipients(channel), GatewayConstant.Opcode.DISPATCH.ordinal(), new ChannelDTO(channel, null), GatewayConstant.Event.CHANNEL_UPDATE.getValue());
+		gatewayService.sendMessageToSpecificSessions(getRecipients(channel), GatewayConstant.Opcode.DISPATCH.ordinal(), new ChannelDTO(channel, null, null, null, null), GatewayConstant.Event.CHANNEL_UPDATE.getValue());
 		log.debug("Channel ({}) edited successfully", channel.getName());
 		return channel;
 	}
@@ -120,7 +136,8 @@ public class ChannelServiceImpl implements ChannelService {
 		Member member = memberService.getByChannelIdAndUserId(channel.getId(), user.getId())
 				.orElseThrow(MemberInChannelNotFoundException::new);
 
-		if (!member.hasAnyPermission(MemberConstant.Permissions.ADMIN)) throw new MissingPermissionsException();
+		if (!member.hasAnyPermission(MemberConstant.Permissions.OWNER, MemberConstant.Permissions.ADMIN))
+			throw new MissingPermissionsException();
 
 		channelRepository.delete(channel);
 		gatewayService.sendMessageToSpecificSessions(getRecipients(channel), GatewayConstant.Opcode.DISPATCH.ordinal(), Map.of("id", channel.getId()), GatewayConstant.Event.CHANNEL_DELETE.getValue());
@@ -133,6 +150,9 @@ public class ChannelServiceImpl implements ChannelService {
 		if (memberService.getByChannelIdAndUserId(channel.getId(), user.getId()).isPresent())
 			throw new MemberAlreadyInChannelException();
 
+		if (channel.getType() == ChannelConstant.Type.DM.getType())
+			throw new ChannelNotFoundException();
+
 		Member member = new Member(user, channel, 0);
 		member.setPermissions(MemberConstant.Permissions.ATTACH_FILES, MemberConstant.Permissions.SEND_MESSAGES);
 		log.debug("Member ({}) joined channel ({}) successfully", member.getUser().getUsername(), channel.getName());
@@ -143,6 +163,9 @@ public class ChannelServiceImpl implements ChannelService {
 	@Override
 	public void removeMember(Channel channel, User user) throws Exception {
 		Member member = memberService.getByChannelIdAndUserId(channel.getId(), user.getId()).orElseThrow(MemberInChannelNotFoundException::new);
+
+		if (channel.getType() == ChannelConstant.Type.DM.getType())
+			throw new ChannelNotFoundException();
 
 		memberService.delete(member);
 		gatewayService.sendMessageToSpecificSessions(getRecipients(channel), GatewayConstant.Opcode.DISPATCH.ordinal(), new MemberDTO(member, true), GatewayConstant.Event.MEMBER_REMOVE.getValue());
